requirements.txt
aiogram>=3.13.1,<4.0.0
python-dotenv>=1.0.1,<2.0.0
